// src/pages/create-project/components/select-project-template/project-template.interface.ts
// INTEL CONFIDENTIAL
//
// Copyright (C) 2021 Intel Corporation
//
// This software and the related documents are Intel copyrighted materials, and your use of them is governed by
// the express license under which they were provided to you ("License"). Unless the License provides otherwise,
// you may not use, modify, copy, publish, distribute, disclose or transmit this software or the related documents
// without Intel's prior written permission.
//
// This software and the related documents are provided as is, with no express or implied warranties,
// other than those that are expressly stated in the License.

import { DOMAIN, SUBDOMAIN } from '../../../../core/projects';
import { LabelsRelationType } from './utils';

export type ProjectTemplateProps = SingleTemplateProps | TaskChainTemplateProps;

export interface SingleTemplateProps {
    cards: DomainCardsMetadata[];
    setSelectedDomains: (domains: DOMAIN[], relations: LabelsRelationType[]) => void;
}

export interface TaskChainTemplateProps {
    subDomains: TaskChainMetadata[];
    selectedDomains: DOMAIN[];
    setSelectedDomains: (domains: DOMAIN[], relations: LabelsRelationType[]) => void;
}
export interface DomainCardsMetadata {
    alt: string;
    id: string;
    domain: DOMAIN;
    subDomain: SUBDOMAIN;
    relation: LabelsRelationType;
    imgSrc: string;
    description: string;
    disabled?: boolean;
}

export interface TaskChainMetadata {
    domains: DOMAIN[];
    relations: LabelsRelationType[];
    description: string;
    images: string[];
}